Move 'receiveFormInfo' into Merlin's connector
