+++ web/app/integration-manager/_demandware-connector/actions.js
@@ -1,0 +1,8 @@
+import {createAction} from '../utils/utils'
+
+// Add other actions here that are specific to this connector.
+// Actions that are returned out of the connector and reduced
+// by the app should go into ./responses.js
+
+// Example.
+// export const myAction = createAction('Some Demandware-specific action')
