+++ web/app/integration-manager/_merlins-connector/actions.js
@@ -1,0 +1,7 @@
+import {createAction} from '../utils/utils'
+
+// Add other actions here that are specific to this connector.
+// Actions that are returned out of the connector and reduced
+// by the app should go into ./responses.js
+
+export const receiveFormInfo = createAction('Receive Form Info')
